skipper: enable skipper_test package
* removes unexported testOptions
* exports run with shutdown channels for tests
