+++ export_test.go
@@ -1,0 +1,3 @@
+package skipper
+
+var RunWithShutdown = run
